Use amazon_link as link to buy individual print copy
[DISCO-336]
