+++ src/app/pages/details/common/get-this-title.js
@@ -4,11 +4,9 @@ import {
     TocOption, WebviewOption, PdfOption, BookshareOption,
     IbooksOption, KindleOption, CheggOption, OptionExpander
 } from './get-this-title-files/options';
-import useAmazonAssociatesLink from '~/pages/details/common/get-this-title-files//amazon-associates-link';
 import OrderPrintCopy from './get-this-title-files/order-print-copy/order-print-copy';
 import './get-this-title-files/get-this-title.scss';
 import trackLink from './track-link';
-import useAmazonIframe from './get-this-title-files/order-print-copy/use-amazon-iframe';
 
 function AdditionalOptions({model}) {
     return (
@@ -31,8 +29,6 @@ export default function GetThisTitle({model}) {
         },
         [model.id]
     );
-    const amazonDataLink = useAmazonAssociatesLink(model.slug);
-    const iframeCode = useAmazonIframe(model.slug);
 
     return (
         <div className="get-the-book">
@@ -49,7 +45,7 @@ export default function GetThisTitle({model}) {
                     <OptionExpander {...{expanded, additionalOptions}} toggle={toggleExpanded} />
                 </div>
             </div>
-            <OrderPrintCopy amazonDataLink={amazonDataLink} iframeCode={iframeCode} />
+            <OrderPrintCopy slug={model.slug} />
         </div>
     );
 }
